agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Agent runtime and individual agents"
requires-python = ">=3.12"
dependencies = [
    "shared",
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "langchain-ollama>=0.2",
    "langchain-google-community[gmail]>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "langfuse>=2.50",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agents"]
